Refactor: reorganize kubernetes.go based on resource types
