+++ test/e2e/kubernetes/serviceaccount.go
@@ -1,0 +1,42 @@
+package kubernetes
+
+import (
+	"context"
+
+	"github.com/go-logr/logr"
+
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/util/retry"
+)
+
+func NewServiceAccount(ctx context.Context, logger logr.Logger, k8s kubernetes.Interface, namespace, name string) error {
+	err := retry.OnError(retry.DefaultRetry, func(err error) bool {
+		// Retry any error type
+		return true
+	}, func() error {
+		if _, err := k8s.CoreV1().ServiceAccounts(namespace).Get(ctx, name, metav1.GetOptions{}); err == nil {
+			logger.Info("Service account already exists", "namespace", namespace, "name", name)
+			return nil
+		}
+
+		serviceAccount := &corev1.ServiceAccount{
+			TypeMeta: metav1.TypeMeta{
+				APIVersion: "v1",
+				Kind:       "ServiceAccount",
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace: namespace,
+				Name:      name,
+			},
+		}
+
+		if _, err := k8s.CoreV1().ServiceAccounts(namespace).Create(ctx, serviceAccount, metav1.CreateOptions{}); err != nil {
+			return err
+		}
+
+		return nil
+	})
+	return err
+}
