+++ test/e2e/kubernetes/version.go
@@ -1,0 +1,29 @@
+package kubernetes
+
+import (
+	"fmt"
+
+	"k8s.io/apimachinery/pkg/util/version"
+)
+
+const (
+	MinimumVersion = "1.25"
+)
+
+func PreviousVersion(kubernetesVersion string) (string, error) {
+	currentVersion, err := version.ParseSemantic(kubernetesVersion + ".0")
+	if err != nil {
+		return "", fmt.Errorf("parsing version: %v", err)
+	}
+	prevVersion := fmt.Sprintf("%d.%d", currentVersion.Major(), currentVersion.Minor()-1)
+	return prevVersion, nil
+}
+
+func IsPreviousVersionSupported(kubernetesVersion string) (bool, error) {
+	prevVersion, err := PreviousVersion(kubernetesVersion)
+	if err != nil {
+		return false, err
+	}
+	minVersion := version.MustParseSemantic(MinimumVersion + ".0")
+	return version.MustParseSemantic(prevVersion + ".0").AtLeast(minVersion), nil
+}
